feat(translation): Report translation results in real-time
Add an `onResult` callback to `translateEpubNodes` to allow for real-time reporting of translation progress and results. This enables the UI to display translated content as it becomes available, improving the user experience for EPUB translations. The `TranslationPage` component is updated to utilize this new callback and add individual translation results to the store.